requirements.txt
mysql.connector
plotly.graph_objects
pandas
matplotlib.pyplot
seaborn
geopandas